Unifying filters into find_ec2_instances()
