+++ src/tests/test_buildercore_core.py
@@ -123,8 +123,8 @@ class SimpleCases(base.BaseCase):
         except core.MultipleRegionsError as e:
             self.assertEqual(["us-east-1", "eu-central-1"], e.regions())
 
-    def test_find_ec2_instance(self):
-        self.assertEquals([], core.find_ec2_instance('dummy1--prod'))
+    def test_find_ec2_instances(self):
+        self.assertEquals([], core.find_ec2_instances('dummy1--prod'))
 
             
 class TestCoreNewProjectData(base.BaseCase):
